Add confirm dialog for deleting component with content, Skip dialog for blank component

+++ src/components/Zone.js
@@ -303,8 +303,24 @@ export class Zone extends React.Component {
   }
 
   removeRow() {
-    const { row, dispatch } = this.props;
-    dispatch(rowActions.removeRow(row.get('id')));
+    const { row, dispatch, zone, html, localState, persistedState } = this.props;
+    const persistedContent = persistedState.get('content');
+    const componentHasNoContent = persistedContent === '<p></p>' || persistedContent === '';
+
+    if (!componentHasNoContent) {
+
+      if (confirm("Are you sure you want to delete this?") == true) {
+        dispatch(rowActions.removeRow(row.get('id')));
+      }
+      else {
+        return false;
+      }
+
+    }
+    else {
+      dispatch(rowActions.removeRow(row.get('id')));
+    }
+
   }
 
   setBoundingBox() {
